requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
asyncpg==0.23.0
databases==0.4.3
jinja2==3.0.2
python-multipart==0.0.16
itsdangerous==2.0.1
bcrypt==3.2.0
python-dotenv==0.19.0